Update_EnemySpawner
오브젝트 풀에 등록되어있는 적 오브젝트를 랜덤으로 생성해주는 기능 추가. => 몹이 한 마리씩 나오는 것이 아닌 동시에 여러마리가 나오는 오류 수정해야함

+++ Assets/Scripts/Player/PlayerAttackController.cs
@@ -7,7 +7,7 @@ public class PlayerAttackController : MonoBehaviour
 {
     [SerializeField] public PlayerSO Data;
 
-    private int curHP;
+    private float curHP;
     private Animator animator;
 
     private bool isAttacking = false;
@@ -86,12 +86,13 @@ public class PlayerAttackController : MonoBehaviour
     {
         isAttacking = true;
         while(enemy!= null && isAttacking)
+        //while (enemy != null )
         {
             enemy.TakeDamage(Data.Damage);
             animator.SetTrigger("Attack");
             yield return new WaitForSeconds(Data.AutoAttackSpeed);
         }
-        isAttacking= false; 
+        isAttacking = false;
     }
 
     public void TakeDamage(int damage)
